ml_pipeline/orchestrator/orchestrator.py
"""
ML Pipeline Orchestrator
========================
Lightweight orchestration layer that coordinates:
1. Earth Engine data export to GCS
2. Vertex AI Custom Training Job
3. Model evaluation and deployment

This script does NOT do heavy ML training - it just tells other services what to do.
"""

import yaml
import logging
import ee
import json
from datetime import datetime, timedelta
from google.cloud import aiplatform, bigquery, storage
import time

logger = logging.getLogger(__name__)

# Load configuration
def load_config():
    """Load configuration from Cloud Storage or local file."""
    try:
        client = storage.Client()
        bucket = client.bucket('carboncheck-data')
        blob = bucket.blob('config/config.yaml')
        yaml_content = blob.download_as_text()
        logger.info("✅ Config loaded from Cloud Storage")
        return yaml.safe_load(yaml_content)
    except Exception as e:
        logger.warning(f"⚠️  Could not load from GCS: {e}")
        with open('config.yaml', 'r') as f:
            logger.info("✅ Config loaded from local file")
            return yaml.safe_load(f)

config = load_config()


def run_pipeline():
    """
    Run the complete ML pipeline.
    
    Returns:
        dict with pipeline results
    """
    start_time = datetime.now()
    
    logger.info("=" * 70)
    logger.info("🌾 ML PIPELINE ORCHESTRATOR")
    logger.info("=" * 70)
    
    try:
        # Step 1: Export Earth Engine data to GCS
        logger.info("STEP 1: Export Earth Engine data to Cloud Storage")
        logger.info("-" * 70)
        export_result = export_earth_engine_data()
        if export_result['status'] != 'success':
            raise Exception(f"Earth Engine export failed: {export_result.get('error')}")
        logger.info(f"✅ Exported {export_result['samples_collected']} samples to GCS")
        logger.info("")
        
        # Step 2: Trigger Vertex AI Training Job
        logger.info("STEP 2: Trigger Vertex AI Custom Training Job")
        logger.info("-" * 70)
        training_result = trigger_training_job()
        if training_result['status'] != 'success':
            raise Exception(f"Training job failed: {training_result.get('error')}")
        logger.info(f"✅ Training complete - Accuracy: {training_result['accuracy']:.2%}")
        logger.info("")
        
        # Step 3: Evaluate and deploy
        logger.info("STEP 3: Evaluate and deploy if quality gates pass")
        logger.info("-" * 70)
        deployment_result = evaluate_and_deploy(training_result)
        if deployment_result['status'] != 'success':
            raise Exception(f"Deployment failed: {deployment_result.get('error')}")
        
        if deployment_result['deployed']:
            logger.info("✅ New model DEPLOYED to production")
        else:
            logger.warning("⛔ New model BLOCKED - did not pass quality gates")
        logger.info("")
        
        # Summary
        duration = (datetime.now() - start_time).total_seconds() / 60
        
        logger.info("=" * 70)
        logger.info("✅ PIPELINE COMPLETE")
        logger.info("=" * 70)
        logger.info(f"Duration: {duration:.1f} minutes")
        logger.info(f"Deployed: {deployment_result['deployed']}")
        
        return {
            'status': 'success',
            'duration_minutes': round(duration, 2),
            'export': export_result,
            'training': training_result,
            'deployment': deployment_result
        }
    
    except Exception as e:
        duration = (datetime.now() - start_time).total_seconds() / 60
        logger.error(f"❌ Pipeline failed: {e}", exc_info=True)
        return {
            'status': 'error',
            'error': str(e),
            'duration_minutes': round(duration, 2)
        }


def export_earth_engine_data():
    """
    Export training data from Earth Engine to Cloud Storage.
    
    Returns:
        dict with export results
    """
    try:
        project_id = config['project']['id']
        bucket_name = config['storage']['bucket']
        samples_per_crop = config['data_collection']['samples_per_crop']
        crops = config['data_collection']['crops']
        
        logger.info(f"   Exporting {samples_per_crop} samples per crop...")
        logger.info(f"   Crops: {[c['name'] for c in crops]}")
        
        # Initialize Earth Engine
        ee.Initialize(project=project_id)
        
        # Export data to GCS as CSV (trainer will load this)
        timestamp = datetime.now().strftime('%Y%m%d_%H%M%S')
        export_path = f'gs://{bucket_name}/training_exports/export_{timestamp}.csv'
        
        # TODO: Implement actual EE export
        # For now, just trigger BigQuery collection
        
        total_samples = samples_per_crop * len(crops)
        
        logger.info(f"   ✅ Would export {total_samples} samples to {export_path}")
        
        return {
            'status': 'success',
            'samples_collected': total_samples,
            'export_path': export_path
        }
    
    except Exception as e:
        logger.error(f"❌ Earth Engine export failed: {e}", exc_info=True)
        return {
            'status': 'error',
            'error': str(e)
        }


def trigger_training_job():
    """
    Trigger Vertex AI Custom Training Job.
    
    Returns:
        dict with training results
    """
    try:
        project_id = config['project']['id']
        region = config['project']['region']
        bucket_name = config['storage']['bucket']
        
        # Training configuration
        machine_type = config.get('training', {}).get('machine_type', 'n1-standard-4')
        
        logger.info(f"   Triggering custom training job...")
        logger.info(f"   Machine type: {machine_type}")
        
        # Initialize Vertex AI
        aiplatform.init(project=project_id, location=region)
        
        # Define custom training job
        job = aiplatform.CustomContainerTrainingJob(
            display_name=f'crop-training-{datetime.now().strftime("%Y%m%d_%H%M")}',
            container_uri=f'{region}-docker.pkg.dev/{project_id}/ml-containers/crop-trainer:latest',
        )
        
        # Run training job
        logger.info("   Starting training job on Vertex AI...")
        model = job.run(
            replica_count=1,
            machine_type=machine_type,
            accelerator_count=0,
            base_output_dir=f'gs://{bucket_name}/training_output'
        )
        
        logger.info("   ✅ Training job complete")
        
        # Get metrics from training output
        # TODO: Load actual metrics from GCS
        
        return {
            'status': 'success',
            'accuracy': 0.85,  # Placeholder
            'job_name': job.display_name
        }
    
    except Exception as e:
        logger.error(f"❌ Training job failed: {e}", exc_info=True)
        return {
            'status': 'error',
            'error': str(e)
        }


def evaluate_and_deploy(training_result):
    """
    Evaluate model and deploy if quality gates pass.
    
    Args:
        training_result: Results from training job
    
    Returns:
        dict with deployment decision
    """
    try:
        gates = config['quality_gates']
        accuracy = training_result['accuracy']
        
        logger.info(f"   Model accuracy: {accuracy:.2%}")
        logger.info(f"   Minimum required: {gates['absolute_min_accuracy']:.0%}")
        
        # Check quality gates
        if accuracy >= gates['absolute_min_accuracy']:
            logger.info("   ✅ Quality gates passed")
            
            # Deploy to Vertex AI endpoint
            project_id = config['project']['id']
            region = config['project']['region']
            bucket_name = config['storage']['bucket']
            
            deploy_model_to_endpoint(project_id, region, bucket_name)
            
            return {
                'status': 'success',
                'deployed': True,
                'accuracy': accuracy
            }
        else:
            logger.warning("   ❌ Quality gates failed")
            return {
                'status': 'success',
                'deployed': False,
                'accuracy': accuracy,
                'reason': f'Accuracy {accuracy:.2%} < minimum {gates["absolute_min_accuracy"]:.0%}'
            }
    
    except Exception as e:
        logger.error(f"❌ Evaluation failed: {e}", exc_info=True)
        return {
            'status': 'error',
            'error': str(e),
            'deployed': False
        }


def deploy_model_to_endpoint(project_id, region, bucket_name):
    """Deploy model to Vertex AI endpoint."""
    aiplatform.init(project=project_id, location=region)
    
    model_name = config['model']['name']
    endpoint_name = config['model']['endpoint_name']
    
    # Upload model
    model = aiplatform.Model.upload(
        display_name=model_name,
        artifact_uri=f'gs://{bucket_name}/models/crop_classifier_latest',
        serving_container_image_uri='us-docker.pkg.dev/vertex-ai/prediction/sklearn-cpu.1-3:latest'
    )
    
    # Get or create endpoint
    existing_endpoints = aiplatform.Endpoint.list(filter=f'display_name="{endpoint_name}"')
    if existing_endpoints:
        endpoint = existing_endpoints[0]
    else:
        endpoint = aiplatform.Endpoint.create(display_name=endpoint_name)
    
    # Undeploy old models
    for deployed_model in endpoint.list_models():
        endpoint.undeploy(deployed_model_id=deployed_model.id)
    
    # Deploy new model
    model.deploy(
        endpoint=endpoint,
        deployed_model_display_name='crop-model-latest',
        machine_type='n1-standard-2',
        min_replica_count=1,
        max_replica_count=3,
        traffic_percentage=100
    )
    
    logger.info(f"   Deployed to endpoint: {endpoint.display_name}")


if __name__ == '__main__':
    # For testing
    result = run_pipeline()
    print(json.dumps(result, indent=2, default=str))

