ml_pipeline/orchestrator/requirements.txt
# Orchestrator Requirements (Lightweight - No ML libraries!)
# =============================================================

# Google Cloud SDK
google-cloud-aiplatform>=1.38.0
google-cloud-bigquery>=3.13.0
google-cloud-storage>=2.10.0
earthengine-api>=0.1.384

# Configuration
PyYAML>=6.0.1

# Web Server
flask>=2.3.0
gunicorn>=21.2.0

